DEV-55 Make submission info available in AutoTest environment (#1361)
This makes some submission information available in the AutoTest environment. Currently this is done by simply setting them as environment variables. It can be disabled on a per-suite basis.

The information is set in the `CG_INFO` environment variable, as a JSON object with the following keys:

* `deadline`: The deadline of this assignment, in ISO 8601 format.
* `submitted_at`: The submission date and time, in ISO 8601 format.
* `result_id`: A unique identifier for this submission.
* `student_id`: A unique identifier for the user who created this submission.

DEV-55 #done

+++ .github/scripts/run_backend_tests.bash
@@ -59,6 +59,9 @@ pytest --cov cg_worker_pool \
        "$(pwd)/cg_cache/tests/" \
        -vvvv
 res1="$?"
+if [[ "$res1" -ne 0 ]]; then
+    exit "$res1";
+fi
 
 timeout -k 600 600 \
         pytest --cov psef --cov cg_signals --cov cg_cache \
@@ -68,6 +71,9 @@ timeout -k 600 600 \
         "$(pwd)/psef_test/test_auto_test.py" \
         -vvvv
 res2="$?"
+if [[ "$res2" -ne 0 ]]; then
+    exit "$res2";
+fi
 
 rm "$(pwd)/psef_test/test_auto_test.py"
 
@@ -80,6 +86,9 @@ timeout -k 900 900 \
         -n 4 \
         -vvvv
 res3="$?"
+if [[ "$res3" -ne 0 ]]; then
+    exit "$res3";
+fi
 
 make doctest
 res4="$?"
